core.luac removed, uncompiled core.lua added.
If you want to test the gamemode you have to download DGS System, create your own images or dont use them, you dont need them to use the gamemode just set the path in the client files to nil and youre good to go.
The database structure is comming soon.